We check output devices everytime we make a call
