+++ src/calls/components/CallButton/CallButtonContainer.js
@@ -1,0 +1,13 @@
+import { connect } from 'react-redux';
+import { CallButtonElement } from './CallButton';
+
+function mapStateToProps({ settings }) {
+  const { devices } = settings;
+  return {
+    devices
+  };
+}
+
+export const CallButton = connect(mapStateToProps)(CallButtonElement);
+
+export default CallButton;
