+++ src/calls/components/UserPhoneNumberButton/UserPhoneNumberButton.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Icon, Button } from 'semantic-ui-react';
+import { Icon } from 'semantic-ui-react';
 import { buildcaller, formatPhoneNumber } from 'calls/utils/utils';
+import { CallButton } from '../CallButton/CallButtonContainer';
 
 export class UserPhoneNumberButton extends Component {
   static propTypes = {
@@ -27,9 +28,15 @@ export class UserPhoneNumberButton extends Component {
 
   render() {
     return (
-      <Button fluid className={"CalleeProfileNumber"} onClick={this.makeCall}>
-        <Icon name={this.props.icon} /> {this.props.phoneNumber}
-      </Button>
+      <CallButton
+        type="modal"
+        content={
+          <>
+            <Icon name={this.props.icon} /> {this.props.phoneNumber}
+          </>
+        }
+        clickHandler={this.makeCall}
+      />
     );
   }
 }
